requirements.txt
python-dotenv
mysql-connector-python
sshtunnel
google-generativeai
paramiko
pymysql